+++ src/Cron/Feed.php
@@ -6,12 +6,7 @@ namespace Omikron\Factfinder\Cron;
 
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Store\Model\StoreManagerInterface;
-use Omikron\Factfinder\Api\Config\ChannelProviderInterface;
-use Omikron\Factfinder\Model\Api\PushImport;
-use Omikron\Factfinder\Model\Export\FeedFactory as FeedGeneratorFactory;
-use Omikron\Factfinder\Model\FtpUploader;
-use Omikron\Factfinder\Model\StoreEmulation;
-use Omikron\Factfinder\Model\Stream\CsvFactory;
+use Omikron\Factfinder\Api\FeedServiceInterface;
 
 class Feed
 {
@@ -20,50 +15,20 @@ class Feed
     /** @var ScopeConfigInterface */
     private $scopeConfig;
 
-    /** @var StoreEmulation */
-    private $storeEmulation;
-
-    /** @var FeedGeneratorFactory */
-    private $feedGeneratorFactory;
-
     /** @var StoreManagerInterface */
     private $storeManager;
 
-    /** @var ChannelProviderInterface */
-    private $channelProvider;
-
-    /** @var CsvFactory */
-    private $csvFactory;
-
-    /** @var FtpUploader */
-    private $ftpUploader;
-
-    /** @var string */
-    private $feedType;
-
-    /** @var PushImport */
-    private $pushImport;
+    /** @var FeedServiceInterface  */
+    private $feedService;
 
     public function __construct(
         ScopeConfigInterface $scopeConfig,
         StoreManagerInterface $storeManager,
-        FeedGeneratorFactory $feedFactory,
-        StoreEmulation $emulation,
-        CsvFactory $csvFactory,
-        FtpUploader $ftpUploader,
-        ChannelProviderInterface $channelProvider,
-        PushImport $pushImport,
-        string $type
+        FeedServiceInterface $feedService
     ) {
-        $this->scopeConfig          = $scopeConfig;
-        $this->storeManager         = $storeManager;
-        $this->feedGeneratorFactory = $feedFactory;
-        $this->storeEmulation       = $emulation;
-        $this->csvFactory           = $csvFactory;
-        $this->ftpUploader          = $ftpUploader;
-        $this->channelProvider      = $channelProvider;
-        $this->pushImport           = $pushImport;
-        $this->feedType             = $type;
+        $this->scopeConfig  = $scopeConfig;
+        $this->storeManager = $storeManager;
+        $this->feedService  = $feedService;
     }
 
     public function execute(): void
@@ -73,15 +38,7 @@ class Feed
         }
 
         foreach ($this->storeManager->getStores() as $store) {
-            $this->storeEmulation->runInStore((int) $store->getId(), function () use ($store) {
-                if ($this->channelProvider->isChannelEnabled((int) $store->getId())) {
-                    $filename = "export.{$this->channelProvider->getChannel()}.csv";
-                    $stream   = $this->csvFactory->create(['filename' => "factfinder/{$filename}"]);
-                    $this->feedGeneratorFactory->create($this->feedType)->generate($stream);
-                    $this->ftpUploader->upload($filename, $stream);
-                    $this->pushImport->execute((int) $store->getId());
-                }
-            });
+            $this->feedService->integrate((int) $store->getId());
         }
     }
 }