+++ src/Model/StoreEmulation.php
@@ -19,11 +19,13 @@ class StoreEmulation
 
     public function runInStore(int $storeId, callable $proceed)
     {
+        $result = null;
         try {
             $this->emulation->startEnvironmentEmulation($storeId, Area::AREA_FRONTEND, true);
-            $proceed();
+            $result = $proceed();
         } finally {
             $this->emulation->stopEnvironmentEmulation();
+            return $result;
         }
     }
 }